Web: Move all nasty JavaScript-in-C code to a separate dedicated file
