+++ src/Http.c
@@ -314,6 +314,8 @@ static void Http_SetRequestHeaders(struct HttpRequest* req) {
 #ifdef CC_BUILD_WEB
 #include <emscripten/emscripten.h>
 #include "Errors.h"
+extern void interop_DownloadAsync(const char* url, int method);
+extern int interop_IsHttpsOnly(void);
 
 cc_bool Http_DescribeError(cc_result res, cc_string* dst) { return false; }
 /* web browsers do caching already, so don't need last modified/etags */
@@ -346,46 +348,12 @@ static void Http_DownloadAsync(struct HttpRequest* req) {
 	String_InitArray(url, urlBuffer);
 	Http_BeginRequest(req, &url);
 	Platform_EncodeUtf8(urlStr, &url);
-
-	/* onFinished = FUNC(data, len, status) */
-	/* onProgress = FUNC(read, total) */
-	EM_ASM_({
-		var url        = UTF8ToString($0);
-		var reqMethod  = $1 == 1 ? 'HEAD' : 'GET';	
-		var onFinished = Module["_Http_OnFinishedAsync"];
-		var onProgress = Module["_Http_OnUpdateProgress"];
-
-		var xhr = new XMLHttpRequest();
-		xhr.open(reqMethod, url);
-		xhr.responseType = 'arraybuffer';
-
-		var getContentLength = function(e) {
-			if (e.total) return e.total;
-
-			try {
-				var len = xhr.getResponseHeader('Content-Length');
-				return parseInt(len, 10);
-			} catch (ex) { return 0; }
-		};
-		
-		xhr.onload = function(e) {
-			var src  = new Uint8Array(xhr.response);
-			var len  = src.byteLength;
-			var data = _malloc(len);
-			HEAPU8.set(src, data);
-			onFinished(data, len || getContentLength(e), xhr.status);
-		};
-		xhr.onerror    = function(e) { onFinished(0, 0, xhr.status);  };
-		xhr.ontimeout  = function(e) { onFinished(0, 0, xhr.status);  };
-		xhr.onprogress = function(e) { onProgress(e.loaded, e.total); };
-
-		try { xhr.send(); } catch (e) { onFinished(0, 0, 0); }
-	}, urlStr, req->requestType);
+	interop_DownloadAsync(urlStr, req->requestType);
 }
 
 static void Http_WorkerInit(void) {
 	/* If this webpage is https://, browsers deny any http:// downloading */
-	httpsOnly = EM_ASM_INT_V({ return location.protocol === 'https:'; });
+	httpsOnly = interop_IsHttpsOnly();
 }
 static void Http_WorkerStart(void) { }
 static void Http_WorkerStop(void)  { }
